CLONING SHAZAM/requirements.txt
streamlit==1.38.0
pandas==2.2.3
sentence-transformers==3.1.1
transformers==4.41.2
chromadb==0.5.5
onnxruntime==1.21.0
openai-whisper==20231117
torch==2.1.2
moviepy==1.0.3
numpy==1.23.5